requirements.txt
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blessed==1.20.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.3
contourpy==1.1.0
croniter==1.3.14
cycler==0.11.0
dateutils==0.6.12
debugpy==1.6.7
decorator==5.1.1
deepdiff==6.3.0
distlib==0.3.6
docker-pycreds==0.4.0
einops==0.6.1
executing==1.2.0
fastapi==0.88.0
filelock==3.9.0
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.4.0
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
huggingface-hub==0.14.1
idna==3.4
imageio==2.31.2
importlib-metadata==6.8.0
importlib-resources==6.0.0
inquirer==3.1.3
ipykernel==6.24.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
lazy_loader==0.3
lightning==2.0.2
lightning-cloud==0.5.34
lightning-utilities==0.8.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mmcv==1.3.11
mpmath==1.2.1
multidict==6.0.4
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.1
omegaconf==2.3.0
opencv-python==4.7.0.72
ordered-set==4.1.0
packaging==23.1
pandas==2.0.1
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.2
prompt-toolkit==3.0.39
protobuf==4.23.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.7
Pygments==2.15.1
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
readchar==4.0.5
requests==2.28.1
rich==13.3.5
scikit-image==0.21.0
scipy==1.11.2
sentry-sdk==1.23.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.22.0
starsessions==1.3.0
sympy==1.11.1
tifffile==2023.8.30
timm==0.6.13
tomli==2.0.1
torch==2.0.1
torchmetrics==1.0.0
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
uvicorn==0.22.0
virtualenv==20.17.1
wandb==0.15.3
wcwidth==0.2.6
websocket-client==1.5.1
websockets==11.0.2
yapf==0.33.0
yarl==1.9.2
zipp==3.16.0
